#5
LevelDirectory is essentially complete!
Return to level select will be added when level select is added.
I consider making LevelDirectory's Directory able to be reordered somehow.
This may be achieve with a "Property Drawer" or an undocumented part of Unity known as "ReorderableList".

Levels are now sorted into Worlds
The first 3 i thought of are
1: Vanilla
2: Spikes
3: Moving platforms
maybe 4 is spikes and platforms?

end of comment