#9
Restart Canvas object is now pinned to the top right of the Canvas, it is no longer a child of the score item. This was originally done to save time when moving the score during level transition, so new code has been added to ScoreFadeIn.cs to move the y position of the restart object along with the Score object.